--- a/src/raftstore/store/keys.rs
+++ b/src/raftstore/store/keys.rs
@@ -81,7 +81,8 @@ pub fn region_raft_prefix(region_id: u64) -> Vec<u8> {
 }
 
 pub fn region_raft_prefix_len() -> usize {
-    REGION_RAFT_PREFIX_KEY.len() + mem::size_of::<u64>()
+    // REGION_RAFT_PREFIX_KEY + region_id + suffix
+    REGION_RAFT_PREFIX_KEY.len() + mem::size_of::<u64>() + 1
 }
 
 pub fn raft_log_key(region_id: u64, log_index: u64) -> Vec<u8> {