raft cf memtable insert with hint prefix extractor.
@siddontang @BusyJay PTAL