@zhangjinpeng1987 2017-02-15
I bench this branch and branch master several times, this pr increase insert qps by 15%-17%, and reduce avg latency by 15%.

this branch
```
1st time
read/write requests:                 1600000 (4455.55 per sec.)
avg:                                            7.32ms
approx.  99 percentile:             13.22ms

2nd time
read/write requests:                 1600000 (4368.64 per sec.)
avg:                                            7.18ms
approx.  99 percentile:             15.55ms
```
master
```
1st time:
read/write requests:                 1600000 (3735.74 per sec.)
avg:                                            8.56ms
approx.  99 percentile:             15.75ms
2nd time:
read/write requests:                 1600000 (3806.85 per sec.)
avg:                                            8.40ms
approx.  99 percentile:             16.29ms
```