@zhangjinpeng1987 2017-02-15
ref https://github.com/pingcap/tikv/issues/1603