@ngaut 2017-02-15
Is the bench result created by sysbench ?