@zhangjinpeng1987 2017-02-16
Yes, the result is created by sysbench. 
When bench random insert, both this branch and master's bottle neck is the cpu of raftstore thread, from the result we get a primary conclusion that this branch reduce about 15% cpu usage of raftstore thread in heavy write case.